// pc/src/services/apis/models/InvestmentOverview.ts
export interface InvestmentOverview {
    /**
     * 
     * @type {UnderConstruction}
     * @memberof InvestmentOverview
     */
    underConstruction: UnderConstruction;
    /**
     * 
     * @type {StatisticalInvestment}
     * @memberof InvestmentOverview
     */
    statisticalInvestment: StatisticalInvestment;
    /**
     * 
     * @type {InvestmentCompletion}
     * @memberof InvestmentOverview
     */
    investmentCompletion: InvestmentCompletion;
    /**
     * 
     * @type {IndustrialChainDistribution}
     * @memberof InvestmentOverview
     */
    industrialChainDistribution: IndustrialChainDistribution;
}